split performance tests into their own executable
